feat: pick the right tau to use

+++ packages/app/modal/compile-circuit.py
@@ -3,7 +3,19 @@ import subprocess
 import os
 app = modal.App()
 
-base_image = modal.Image.from_registry("javiersuweijie/zk-email-sdk-base:0.0.1", add_python="3.11")
+base_image = (
+    modal.Image.from_registry("javiersuweijie/zk-email-sdk-base:0.0.1", add_python="3.11")
+    #install curl
+    .run_commands(
+        "apt-get update && apt-get install -y curl"
+    )
+    .run_commands(
+        "curl -o /powersOfTau28_hez_final_24.ptau https://storage.googleapis.com/zkevm/ptau/powersOfTau28_hez_final_24.ptau"
+    )
+    .run_commands(
+        "curl -o /powersOfTau28_hez_final_22.ptau https://storage.googleapis.com/zkevm/ptau/powersOfTau28_hez_final_22.ptau"
+    )
+)
 secret = modal.Secret.from_name(
     "google-storage",
     required_keys=["GOOGLE_ACCESS_KEY_ID", "GOOGLE_ACCESS_KEY_SECRET"]
@@ -40,6 +52,28 @@ def compile_circuit(slug: str):
         circuitPath = circuit_files[0]
     circuitName = circuitPath.split("/")[-1].split(".")[0]
     subprocess.run(["mkdir", "-p", f"/temp_output/{slug}"], check=True)
-    subprocess.run(["circom", circuitPath, "--r1cs", "--wasm", "-o", f"/temp_output/{slug}", "-l", "/project/node_modules"], check=True)
-    subprocess.run(f"NODE_OPTIONS='--max-old-space-size=16000' snarkjs zkey new /temp_output/{slug}/{circuitName}.r1cs /powersOfTau28_hez_final_23.ptau /temp_output/{slug}/{circuitName}.zkey -v", check=True, shell=True)
+    
+    result = subprocess.run(["circom", circuitPath, "--r1cs", "--wasm", "-o", f"/temp_output/{slug}", "-l", "/project/node_modules"], check=True, capture_output=True, text=True)
+    print(result.stdout)
+
+    # find the number of non-linear constraints to determine the power of tau
+    non_linear_constraints_match = re.search(r'non-linear constraints: (\d+)', result.stdout)
+    if non_linear_constraints_match:
+        non_linear_constraints = int(non_linear_constraints_match.group(1))
+        print(f"Number of non-linear constraints: {non_linear_constraints}")
+    else:
+        print("Could not find non-linear constraints in the output")
+    power = math.max(math.floor(math.log(non_linear_constraints, 2)), 22)
+
+    # check if the power of tau exists
+    if not os.path.exists(f"/powersOfTau28_hez_final_{power}.ptau"):
+        print(f"Tau of power {power} not found, downloading...")
+        subprocess.run(f"curl -o /powersOfTau28_hez_final_{power}.ptau https://storage.googleapis.com/zkevm/ptau/powersOfTau28_hez_final_{power}.ptau", check=True, shell=True)
+
+    # generate the zkey
+    subprocess.run(f"NODE_OPTIONS='--max-old-space-size=32768' snarkjs zkey new /temp_output/{slug}/{circuitName}.r1cs /powersOfTau28_hez_final_{power}.ptau /temp_output/{slug}/{circuitName}.zkey -v", check=True, shell=True)
+
+    # check if the zkey was created
+    if not os.path.exists(f"/temp_output/{slug}/{circuitName}.zkey") or os.path.getsize(f"/temp_output/{slug}/{circuitName}.zkey") == 0:
+        raise Exception(f"zkey {circuitName} was not created")
     subprocess.run([f"cp -R /temp_output/* /output/circuit/"], check=True, shell=True)
